feat(docs): migrate Sidebar
