+++ src/components/Sidebar/Sidebar.docs.mdx
@@ -1,27 +1,15 @@
----
-name: Sidebar
-route: /components/sidebar
-menu: Components
----
-
-import { Playground, Props } from 'docz';
-import { Fragment } from 'react';
-import Sidebar from '../../../src/components/Sidebar';
-import Status from '../../utils/Statuses';
-import Text from '../../../src/components/Text';
-import SidebarWithState from './sidebar';
+import { Status, Props, Story } from '../../../docs/components';
+import Sidebar from './Sidebar';
 
 # Sidebar
 
-<Status.InReview/> <Status.Description size={Text.KILO}>Under active development</Status.Description>
+<Status.InReview/><Status.Description>Under active development</Status.Description>
 
 The sidebar is the primary navigational component on SumUp's web-based applications. It groups different
 sections of the same product (the merchant dashboard, for example), allowing easy navigation between first
 and the second level of navigations, keeping a strong anchor for our merchants to always know where they are.
 
-<Playground>
-  <SidebarWithState />
-</Playground>
+<Story id="components-sidebar--base" />
 
 <Props of={Sidebar} />
 
@@ -43,7 +31,6 @@ which will act as an aggregator.
 - **Do** group related items of a given action under a secondary
 - **Do not** place any element or part of an element on the sidebar.
 - **Do not** use the secondary level as anchors of the first level.
-
 
 ## Content guidelines
 
@@ -71,7 +58,7 @@ Sidebar:
 When you want to create a sidebar with a flat navigation structure, you
 can simply use something like this:
 
-```
+```jsx
 import React, { useState } from 'react';
 import { Sidebar } from '@sumup/circuit-ui';
 
@@ -110,18 +97,10 @@ In order to create submenus, you need to wrap the `Sidebar.NavItem` components
 that you want to put inside a parent menu with the `Sidebar.Aggregator`
 component. You can do that like this:
 
-```
+```jsx
 <Sidebar.Aggregator label="Parent">
-  <Sidebar.NavItem
-    onClick={() => null}
-    label="Child 1"
-    selected={true}
-  />
-  <Sidebar.NavItem
-    onClick={() => null}
-    label="Child 2"
-    selected={false}
-  />
+  <Sidebar.NavItem onClick={() => null} label="Child 1" selected={true} />
+  <Sidebar.NavItem onClick={() => null} label="Child 2" selected={false} />
 </Sidebar.Aggregator>
 ```
 
@@ -131,7 +110,7 @@ To avoid keeping the logic of toggling the `Sidebar` on your application, you ca
 `SidebarContextProvider` and `SidebarContextConsumer` that are implemented with the new version of
 React's Context API. You can do that like this:
 
-```
+```jsx
 import React, { useState } from 'react';
 import { Sidebar, SidebarContextProvider, SidebarContextConsumer } from '@sumup/circuit-ui';
